+++ packages/universal-wallet-react/src/molecules/dialogs/AddDialog/index.tsx
@@ -1,1 +1,2 @@
 export * from './AddDialogContent';
+export * from './buildAddInterface';
