+++ packages/universal-wallet-react/src/molecules/dialogs/ExploreDialog/index.tsx
@@ -1,1 +1,2 @@
 export * from './ExploreDialogContent';
+export * from './buildExploreInterface';
