+++ packages/universal-wallet-react/src/molecules/dialogs/ExportDialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './buildExportInterface';
