+++ packages/universal-wallet-react/src/molecules/dialogs/GenerateDialog/index.tsx
@@ -1,0 +1,1 @@
+export * from './buildGenerateInterface';
