+++ packages/universal-wallet-react/src/molecules/dialogs/RemoveDialog/index.tsx
@@ -1,1 +1,2 @@
 export * from './RemoveDialogContent';
+export * from './buildRemoveInterface';
