+++ packages/universal-wallet-react/src/plugins/did-key/DidKeyWallet.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Meta, Story } from '@storybook/react';
+
+import { DidKeyWallet as Component, Props } from './DidKeyWallet';
+
+const meta: Meta = {
+  title: 'Plugins/DID Key',
+  component: Component,
+};
+
+export default meta;
+
+const Template: Story<Props> = (args) => <Component {...args} />;
+
+export const DIDKey = Template.bind({});
+
+DIDKey.args = {};
